Refine caller opt-in to vartime implementations
Remove SetVarTime from the group interface, to make it lighter.

Create a new interface to allow callers to detect if vartime
is available and use it for public data if they so choose.

In group/edwards25519, make the existing vartime algorithm
correctly built only with build tag vartime.

Closes #250.
