+++ CHANGELOG
@@ -37,9 +37,7 @@ v0 to v1:
 
     - Changed order of arguments for `Point.Mul()`. It now follows the
       mathematical additive notation with the scalar in front:
-        -> `Mul(kyber.Scalar, kyber.Point) kyber.Point`. 
-    - Added `SetVarTime` to `Scalar` and `Point` that may use a variable time
-      implementation if available.
+        -> `Mul(kyber.Scalar, kyber.Point) kyber.Point`.
     - commented out the dh_test.go which is not up-to-date anymore
     - taken onet/crypto/key{_test}.go to kyber/util/encoding
         + Removed abstract.Suite, only taking kyber.Group now
