+++ group/edwards25519/allow_test.go
@@ -1,0 +1,16 @@
+// +build !vartime
+
+package edwards25519
+
+import (
+	"testing"
+
+	"github.com/dedis/kyber"
+)
+
+func TestNotVartime(t *testing.T) {
+	p := testSuite.Point()
+	if _, ok := p.(kyber.AllowsVarTime); ok {
+		t.Fatal("expected Point to NOT allow var time")
+	}
+}
