+++ group/edwards25519/allowvt_test.go
@@ -1,0 +1,22 @@
+// +build vartime
+
+package edwards25519
+
+import (
+	"testing"
+
+	"github.com/dedis/kyber"
+)
+
+func TestVartime(t *testing.T) {
+	p := testSuite.Point()
+	if pvt, ok := p.(kyber.AllowsVarTime); ok {
+		// Try both settings
+		pvt.AllowVarTime(false)
+		p.Mul(one, p)
+		pvt.AllowVarTime(true)
+		p.Mul(one, p)
+	} else {
+		t.Fatal("expected Point to allow var time")
+	}
+}
